# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for collecting GitHub stats.",
        "title": "Github Harvester API",
        "contact": {}
    },
    "paths": {
        "/repo/{owner}/{repo}": {
            "get": {
                "description": "Returns stars, forks, and commit count for a public repo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Get GitHub Repository Stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository Owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository Name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RepositoryModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.RepositoryModel": {
            "type": "object",
            "properties": {
                "commitsCount": {
                    "type": "integer",
                    "format": "int64"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forks": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    }
}